# docs/swagger.yaml
basePath: /
definitions:
...
  input.RoleAddRoleUsersReq:
    properties:
      adminIds:
        description: 管理员ID（支持多个）
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      roleId:
        description: 角色ID
        example: 2
        type: integer
    required:
    - adminIds
    - roleId
    type: object
  input.RoleDeleteReq:
    properties:
      id:
        description: 角色ID
        example: 2
        type: integer
    required:
    - id
    type: object
  input.RoleListPageReq:
    properties:
      enable:
        description: 状态
        enum:
        - 0
        - 1
        example: 1
        type: integer
      name:
        description: 名称，支持模糊搜索
        example: 超级
        type: string
      pageNo:
        description: 页码
        example: 1
        type: integer
      pageSize:
        description: 每页展示条数
        example: 20
        type: integer
    required:
    - pageNo
    - pageSize
    type: object
  input.RoleRemoveRoleUsersReq:
    properties:
      adminIds:
        description: 管理员ID（支持多个）
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      roleId:
        description: 角色ID
        example: 2
        type: integer
    required:
    - adminIds
    - roleId
    type: object
  input.RoleSaveReq:
    properties:
      code:
        description: 角色标识
        example: TestAdmin
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      id:
        description: 角色ID
        example: 2
        type: integer
      name:
        description: 角色名称
        example: 测试管理员
        type: string
      permissionIds:
        description: 角色绑定菜单ID
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
...
  resp.RoleListAllItem:
    properties:
      code:
        description: 角色标识
        example: SUPER_ADMIN
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      id:
        description: 角色ID
        example: 1
        type: integer
      name:
        description: 角色名称
        example: 超级管理员
        type: string
    type: object
  resp.RoleListAllResp:
    properties:
      list:
        description: 角色列表
        items:
          $ref: '#/definitions/resp.RoleListAllItem'
        type: array
    type: object
  resp.RoleListPageItem:
    properties:
      code:
        description: 角色标识
        example: SUPER_ADMIN
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      id:
        description: 角色ID
        example: 1
        type: integer
      name:
        description: 角色名称
        example: 超级管理员
        type: string
      permissionIds:
        description: 角色绑定菜单ID
        items:
          type: integer
        type: array
    type: object
  resp.RoleListPageResp:
    properties:
      pageData:
        description: 当前页码数据
        items:
          $ref: '#/definitions/resp.RoleListPageItem'
        type: array
      total:
        description: 总计条数
        example: 100
        type: integer
    type: object
  resp.RoleMenuItem:
    properties:
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/resp.RoleMenuItem'
        type: array
      code:
        description: 菜单标识
        example: Base
        type: string
      component:
        description: 组件路径
        example: /src/list/list.vue
        type: string
      description:
        description: 描述
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      icon:
        description: 菜单图标
        example: i-fe:list
        type: string
      id:
        description: 菜单ID
        example: 1
        type: integer
      keepAlive:
        description: 保持活跃
        example: false
        type: boolean
      layout:
        description: 布局
        example: full
        type: string
      method:
        description: 方法
        type: string
      name:
        description: 菜单名称
        example: 基础菜单
        type: string
      order:
        description: 排序（从小到大）
        example: 0
        type: integer
      parentId:
        description: 父级ID
        example: 0
        type: integer
      path:
        description: 菜单路径
        example: /path/url
        type: string
      redirect:
        description: 重定向
        type: string
      show:
        description: 显示状态
        example: true
        type: boolean
      type:
        description: 菜单类型
        enum:
        - BUTTON
        - MENU
        example: MENU
        type: string
    type: object
  resp.RolePermissionsResp:
    properties:
      menu:
        items:
          $ref: '#/definitions/resp.RoleMenuItem'
        type: array
    type: object
  response.Response:
    properties:
      code:
        description: 响应code
        example: 0
        type: integer
      data:
        description: 响应数据
      msg:
        description: 响应信息
        example: success
        type: string
    type: object
info:
  contact:
    email: junjie.he.925@gmail.com
    name: API 支持
  description: OneAdmin 系统接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: OneAdmin API
  version: "1.0"
paths:
...
  /api/admin/roles/add-role-users:
    post:
      description: 将指定角色批量分配给管理员（建立管理员与角色的关联关系）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 分配参数（角色ID + 管理员ID列表）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.RoleAddRoleUsersReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 为角色分配管理员
      tags:
      - 角色
  /api/admin/roles/all:
    post:
      description: 获取系统所有角色的简要信息（不分页），用于下拉选择或管理员绑定角色
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RoleListAllResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取全部角色列表
      tags:
      - 角色
  /api/admin/roles/delete:
    post:
      description: 根据角色 ID 删除角色（通常不允许删除已绑定管理员或正在使用的角色）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 删除参数（角色 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.RoleDeleteReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色
  /api/admin/roles/list:
    post:
      description: 获取角色分页列表，支持按角色名称模糊搜索及状态筛选，用于后台角色管理与权限分配
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 分页查询参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.RoleListPageReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RoleListPageResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取角色列表（分页）
      tags:
      - 角色
  /api/admin/roles/permissions:
    post:
      description: 获取当前登录管理员在当前角色下的菜单及按钮权限树（已按权限过滤），用于前端动态路由及权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RolePermissionsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取当前管理员权限菜单树
      tags:
      - 角色
  /api/admin/roles/remove-role-users:
    post:
      description: 将指定管理员从角色中批量移除（解除管理员与角色的关联关系）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 移除参数（角色ID + 管理员ID列表）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.RoleRemoveRoleUsersReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 从角色移除管理员
      tags:
      - 角色
  /api/admin/roles/save:
    post:
      description: 根据是否传入 ID 创建或更新角色信息，同时可绑定菜单及按钮权限（权限ID集合）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 角色保存参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.RoleSaveReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建或更新角色
      tags:
      - 角色
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"